Fix houdini support
Installing an apk works in the following way through play store. First play store checks the abi and abi2 in build.prop. If the apk is available it checks again while installing the apk if the apk is for a specific arch. First the abi(x86), then the abi2(armeabi-v7a, the fastest one) and if even that one doesnt work the upgradeabi(armeabi, a little slower). 

Most apk's will work with the first and second, but if it doesn't, fall back to the third. Houdini supports armeabi, armeabi-v7a, armeabi-v7a neon.

Still testing and I maybe incorrect