smi: redefine define audioserver
N brought in this change https://goo.gl/R09n2S
This seperates audioserver from mediaserver process.
We had previously redefined mediaserver to add system & radio to the
group for the intel modem. With audioserver now it's own service we
need to add radio & system to audio server as well.

This gives modem access to the needed audio ports for in-call audio.

Change-Id: I8c52e0b5f5c768d8d370f8c9ec4f2bc75175a9cc